Disable deploy step for releases
Whenever the version is set to a release version (like during a
release), the deploy step on the master branch will fail as the version
is not a snapshot (and is attempting to deploy to the snapshots
repository). This refactors some tool usage and checks the maven project
version to prevent this error.